@@ -1,8 +1,8 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+
     <DockerComposeProjectPath>..\..\docker-compose.dcproj</DockerComposeProjectPath>
     <UserSecretsId>42811497-7821-49c9-b654-b7ef751e50ec</UserSecretsId>
   </PropertyGroup>
@@ -14,19 +14,20 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="Autofac.Extensions.DependencyInjection" Version="4.4.0" />
-    <PackageReference Include="MediatR" Version="7.0.0" />
-    <PackageReference Include="MediatR.Extensions.Microsoft.DependencyInjection" Version="7.0.0" />
-    <PackageReference Include="Microsoft.AspNetCore.App" />
-    <PackageReference Include="Swashbuckle.AspNetCore" Version="4.0.1" />
+    <PackageReference Include="Autofac.Extensions.DependencyInjection" Version="6.0.0" />
+    <PackageReference Include="MassTransit" Version="7.0.2" />
+    <PackageReference Include="MediatR" Version="8.0.0" />
+    <PackageReference Include="MediatR.Extensions.Microsoft.DependencyInjection" Version="8.0.0" />
+    <PackageReference Include="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="3.1.2" />
+    <PackageReference Include="Swashbuckle.AspNetCore" Version="5.0.0" />
   </ItemGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\..\src\Domain.Core\Domain.Core.csproj" />
-    <ProjectReference Include="..\..\src\Infrastructure.Crosscutting.EventBus.Cap\Infrastructure.Crosscutting.EventBus.Cap.csproj" />
     <ProjectReference Include="..\..\src\Infrastructure.Crosscutting.Json.NewtonsoftJson\Infrastructure.Crosscutting.Json.NewtonsoftJson.csproj" />
     <ProjectReference Include="..\..\src\Infrastructure.Crosscutting\Infrastructure.Crosscutting.csproj" />
     <ProjectReference Include="..\..\src\Infrastructure.Data.MongoDB\Infrastructure.Data.MongoDB.csproj" />
+    <ProjectReference Include="..\..\src\Infrastructure.IntegrationEventService\Infrastructure.IntegrationEventService.csproj" />
     <ProjectReference Include="..\Catalog.Domain\Catalog.Domain.csproj" />
     <ProjectReference Include="..\CataLog.Infrastructure.MongoDB.Repository\CataLog.Infrastructure.MongoDB.Repository.csproj" />
   </ItemGroup>
