@@ -4,7 +4,6 @@ using System.Linq;
 using System.Threading.Tasks;
 using Autofac;
 using MySvc.DotNetCore.Framework.Infrastructure.Crosscutting.EventBus;
-using MySvc.DotNetCore.Framework.Infrastructure.Crosscutting.EventBus.Cap;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
@@ -15,7 +14,7 @@ namespace Catalog.API.Extensions
     {
         public static IServiceCollection AddEventBus(this IServiceCollection services, IConfiguration configuration)
         {
-            services.AddSingleton<IEventBus, CapEventBus>();
+            //services.AddSingleton<IEventBus, CapEventBus>();
             //services.AddTransient<OrderStatusChangedToAwaitingValidationIntegrationEventHandler>();
             //services.AddTransient<OrderStatusChangedToPaidIntegrationEventHandler>();
 
