@@ -1,11 +1,11 @@
-﻿using System;
+﻿using MySvc.DotNetCore.Framework.Domain.Core.DomainEvents;
+using MySvc.DotNetCore.Framework.Infrastructure.Crosscutting.EventBus;
+using Sample.Order.Domain.Events;
+using System;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using MySvc.DotNetCore.Framework.Domain.Core.DomainEvents;
-using MySvc.DotNetCore.Framework.Infrastructure.Crosscutting.EventBus;
-using Sample.Order.Application.IntegrationEvents.Events;
-using Sample.Order.Domain.Events;
+using Contracts.Events;
 
 namespace Sample.Order.Application.DomainEventHandlers
 {
@@ -14,28 +14,34 @@ namespace Sample.Order.Application.DomainEventHandlers
     /// </summary>
     public class OrderCreatedDomainEventHandler : IDomainEventHandler<OrderCreatedDomainEvent>
     {
-        private readonly IEventBus _eventBus;
+        private readonly IIntegrationEventService _integrationEventService;
 
-        public OrderCreatedDomainEventHandler(IEventBus eventBus)
+        public OrderCreatedDomainEventHandler(IIntegrationEventService integrationEventService)
         {
-            
-            _eventBus = eventBus ?? throw new ArgumentNullException(nameof(eventBus));
+
+            _integrationEventService = integrationEventService ?? throw new ArgumentNullException(nameof(integrationEventService));
         }
-        
+
         public async Task Handle(OrderCreatedDomainEvent domainEvent, CancellationToken cancellationToken)
         {
             var order = domainEvent.Order;
             //处理领域相关的事情
 
             //保存集成事件
-            await _eventBus.Publish(
-                new OrderCreatedIntegrationEvent(order.OrderNo, order.OrderTime, 
-                    order.OrderItems.Select(x => new OrderItem()
-                    {
-                        SKU = x.Product.SKU,
-                        Title = x.Product.Title,
-                        Units = x.Units
-                    }).ToList()));
+
+            var orderItems = order.OrderItems.Select(x => new OrderItem
+            {
+                SKU = x.Product.SKU,
+                Title = x.Product.Title,
+                Units = x.Units
+            }).ToArray();
+            await _integrationEventService.SaveIntegrationEvent<OrderCreatedIntegrationEvent>(
+                new OrderCreatedIntegrationEvent
+                {
+                    OrderNo = order.OrderNo,
+                    OrderTime = order.OrderTime,
+                    OrderItems = orderItems
+                });
         }
     }
 }
