@@ -1,12 +1,5 @@
 ﻿using Autofac;
-using MySvc.DotNetCore.Framework.Domain.Core;
-using MySvc.DotNetCore.Framework.Infrastructure.Adapter.AutoMapper;
-using MySvc.DotNetCore.Framework.Infrastructure.Crosscutting.Adapter;
-using MySvc.DotNetCore.Framework.Infrastructure.Crosscutting.EventBus;
-using MySvc.DotNetCore.Framework.Infrastructure.Crosscutting.EventBus.Cap;
 using MySvc.DotNetCore.Framework.Infrastructure.Crosscutting.Json;
-using MySvc.DotNetCore.Framework.Infrastructure.Data.MongoDB;
-using MySvc.DotNetCore.Framework.Infrastructure.Data.MongoDB.Impl;
 using MySvc.DotNetCore.Framework.Infrastructure.Json.NewtonsoftJson;
 
 namespace Sample.Product.Api.DI.AutofacModules
@@ -16,17 +9,16 @@ namespace Sample.Product.Api.DI.AutofacModules
     /// </summary>
     public class CommonModule : Module
     {
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="builder"></param>
         protected override void Load(ContainerBuilder builder)
         {
             //JSON转换器
             builder.RegisterType<NewtonsoftJsonConverter>().As<IJsonConverter>().SingleInstance();
 
-            //注册类型适配转换器
-            builder.RegisterType<AutomapperTypeAdapter>().As<ITypeAdapter>().SingleInstance();
 
-
-            builder.RegisterType<CapEventBus>().As<IEventBus>()
-                .InstancePerLifetimeScope();
         }
     }
 }
