@@ -1,32 +1,44 @@
 ﻿using System;
+using Autofac.Extensions.DependencyInjection;
 using Microsoft.AspNetCore;
 using Microsoft.AspNetCore.Hosting;
+using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using NLog.Web;
 using Sample.Product.Api.Extensions;
 
 namespace Sample.Product.Api
 {
+    /// <summary>
+    /// 
+    /// </summary>
     public class Program
     {
-        public static void Main(string[] args)
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="args"></param>
+        /// <returns></returns>
+        public static int Main(string[] args)
         {
             var logger = NLog.Web.NLogBuilder.ConfigureNLog("nlog.config").GetCurrentClassLogger();
 
             try
             {
                 logger.Info("init main !~~");
-                CreateWebHostBuilder(args)
+                CreateHostBuilder(args)
                     .Build()
                     .MigrateMongoDB() //初始化数据库
                     .Run();
+                return 0;
 
             }
             catch (Exception ex)
             {
                 //NLog: catch setup errors
                 logger.Error(ex, "Stopped program because of exception");
-                throw;
+                return 1;
+
             }
             finally
             {
@@ -35,14 +47,23 @@ namespace Sample.Product.Api
             }
         }
 
-        public static IWebHostBuilder CreateWebHostBuilder(string[] args) =>
-            WebHost.CreateDefaultBuilder(args)
-                .UseStartup<Startup>()
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="args"></param>
+        /// <returns></returns>
+        public static IHostBuilder CreateHostBuilder(string[] args) =>
+            Host.CreateDefaultBuilder(args)
+                .UseServiceProviderFactory(new AutofacServiceProviderFactory())
                 .ConfigureLogging(logging =>
                 {
                     logging.ClearProviders();
                     logging.SetMinimumLevel(LogLevel.Information);
                 })
-                .UseNLog();
+                .UseNLog()
+                .ConfigureWebHostDefaults(webBuilder =>
+                {
+                    webBuilder.UseStartup<Startup>();
+                });
     }
 }
