@@ -2,8 +2,11 @@ using MediatR;
 
 namespace MySvc.DotNetCore.Framework.Domain.Core.DomainEvents
 {
+    /// <summary>
+    /// �����¼��ӿ�
+    /// </summary>
     public interface IDomainEvent : INotification
     {
-        
+
     }
 }
