@@ -1,1 +1,62 @@
-﻿
+﻿//=======================================================================================
+//
+// Copyright (C) 2010-2011 Asia-Peak Technologies, Inc. All Rights Reserved.
+// 
+// All the code, text, graphics, media, design, programs and other works are
+// protected by copyright law. Unauthorized reproduction or redistribution of them,  
+// or any portion of them, are forbidden.
+// 
+//=======================================================================================
+//---------------------------------------------------------------------------------------
+// <auto-generated>
+//     This code was generated from a template.
+//
+//     Changes to this file may cause incorrect behavior and will be lost if
+//     the code is regenerated.
+// </auto-generated>
+//---------------------------------------------------------------------------------------
+
+using System;
+
+namespace MySvc.DotNetCore.Framework.Infrastructure.Authorization.Admin.Permissions
+{	
+
+
+	/// <summary>
+    /// 权限码定义
+    /// </summary>
+	public static class Permission
+    {
+		
+		/// <summary>
+		/// 账号管理
+		/// </summary>
+		public const string AccountManage = "AccountManage";
+		
+		/// <summary>
+		/// 员工列表
+		/// </summary>
+		public const string UserList = "UserList";
+		
+		/// <summary>
+		/// 添加员工
+		/// </summary>
+		public const string AddUser = "AddUser";
+		
+		/// <summary>
+		/// 查看员工资料
+		/// </summary>
+		public const string ViewUser = "ViewUser";
+		
+		/// <summary>
+		/// 编辑员工资料
+		/// </summary>
+		public const string EditUser = "EditUser";
+		
+		/// <summary>
+		/// 删除员工
+		/// </summary>
+		public const string DeleteUser = "DeleteUser";
+		
+	}
+}
