@@ -3,7 +3,6 @@ using Microsoft.Extensions.Caching.Distributed;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using Microsoft.Extensions.Primitives;
-using MySvc.DotNetCore.Framework.Infrastructure.Authorization.Admin.Exceptions;
 using MySvc.DotNetCore.Framework.Infrastructure.Authorization.Client.Extensions;
 using MySvc.DotNetCore.Framework.Infrastructure.Crosscutting.Helpers;
 using MySvc.DotNetCore.Framework.Infrastructure.Crosscutting.Json;
@@ -30,7 +29,7 @@ namespace MySvc.DotNetCore.Framework.Infrastructure.Authorization.Client
             IJsonConverter jsonConverter,
             IHttpClientFactory httpClientFactory,
             ILogger<UserIdentityService> logger,
-            
+
             IOptions<AuthServiceOptions> authServiceOptionsAccessor)
         {
             _contextAccessor = contextAccessor ?? throw new ArgumentNullException(nameof(contextAccessor));
@@ -48,14 +47,14 @@ namespace MySvc.DotNetCore.Framework.Infrastructure.Authorization.Client
             {
                 throw new AuthenticationException("unauthorized");
             }
-            
+
             var tenantUser = MapTenantUser();
             return tenantUser;
         }
 
         public async Task<UserIdentity> GetUserIdentityAsync()
         {
-            
+
             if (_contextAccessor.HttpContext == null)
             {
                 _logger.LogError("_contextAccessor.HttpContext is null");
@@ -114,7 +113,7 @@ namespace MySvc.DotNetCore.Framework.Infrastructure.Authorization.Client
                 }
                 catch (Exception ex)
                 {
-                    _logger.LogError(ex,"");
+                    _logger.LogError(ex, "");
                     throw new AuthValidationError(Error.Codes.RequestUserProfileFailed, Error.Names.RequestUserProfileFailed, ex);
                 }
             }
@@ -154,12 +153,12 @@ namespace MySvc.DotNetCore.Framework.Infrastructure.Authorization.Client
             bool.TryParse(email_verified, out bool_email_verified);
 
             bool bool_phone_number_verified = false;
-            bool.TryParse(phone_number_verified, out  bool_phone_number_verified);
+            bool.TryParse(phone_number_verified, out bool_phone_number_verified);
 
             bool bool_hasPaymentPassword = false;
             bool.TryParse(hasPaymentPassword, out bool_hasPaymentPassword);
-            
-            var userIdentity =  new UserIdentity(tenantUserId, tenantCode, userName, fullName,
+
+            var userIdentity = new UserIdentity(tenantUserId, tenantCode, userName, fullName,
                 email, bool_email_verified, dialcode, phone_number, bool_phone_number_verified, bool_hasPaymentPassword, role, new List<string>());
 
             _logger.LogDebug(_jsonConverter.SerializeObject(userIdentity));
